# ENH: Faster image stack navigation