--- a/hyperspy/drawing/image.py
+++ b/hyperspy/drawing/image.py
@@ -231,7 +231,7 @@ class ImagePlot(BlittedFigure):
             am = stack_iterable
         else:
             # Only plot current frame
-            stack_iterable = [self.axes_manager.indices]
+            stack_iterable = [0]
             am = self.axes_manager
 
         if self._text is not None: